Filesystem/Terminal improvements; userspace programs
Changelog:
- set fixed block size of 4KiB for ext2 filesystem
- removed obsolete StdioDevice
- implement file system cache
	- handles accesses to blocks on the file system
	- returns refcounted fs_block_t structs with a pointer to the
	  mapped block
	- block is flushed to disk when the last reference drops
	  (or forcibly flushed with sync())
	- all block-level accesses go through this cache
	- currently no *actual* caching is done (just refcounting)
- implemented getdents syscall (similar to Linux syscall)
- implemented dirent.h methods (fd)opendir, readdir, closedir
- implemented handling of e0 scancodes for PS2KeyboardDevice
- implemented write() for existing files on the ext2 filesystem
- implemented path parsing according to POSIX standard
- implemented traversal for softlinks
- implemented current working directory for processes
- extended AVLTree implementation with emplace method
- extended LibK::string implementation with += operators, insert
  and erase methods
- consolidated read and write methods for BlockDevices with a single
  transfer method
- extended Console API with specialized write functions
- Console now exposes cell_t type
- implemented various xterm control sequences:
  alternate buffer, cursor saving, arrow keys, cursor movement, better
  scrolling, character repeating
- implemented basic cat program
- started implementing basic vi-like editor
	- missing some control sequences ncurses uses
- implemented basic ls program
- fixed backspace handling for sh
- added test/test.txt with Lorem Ipsum for testing the editor

+++ src/userland/libc/sys/types.h
@@ -11,7 +11,7 @@ typedef int blkcnt_t;
 typedef int blksize_t;
 typedef int dev_t;
 typedef int gid_t;
-typedef int ino_t;
+typedef size_t ino_t;
 typedef unsigned mode_t;
 typedef int nlink_t;
 typedef size_t off_t;
